requirements.txt
nltk>=3.8.1
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
clip>=1.0
rembg>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
psutil>=5.9.0 
requests>=2.28.0
sentence-transformers>=2.2.0
easyocr>=1.7.0
openai>=1.3.0
python-dotenv>=1.0.0
moondream>=0.1.7
flask>=2.3.0
flask-cors>=4.0.0
# Text processing dependencies
pdfplumber>=0.9.0
pypdf>=3.0.0
python-pptx>=0.6.0
openpyxl>=3.1.0
# Additional dependencies (optional, for command line fallbacks)
# antiword (install via system package manager)
# libreoffice (install via system package manager) 
# pandoc (install via system package manager)
# poppler-utils (install via system package manager) 
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.8
pillow==10.1.0
easyocr==1.7.0
moondream==0.1.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
numpy==1.25.2 